requirements.txt
attrs==17.2.0
arrow==0.12.0
azure-storage-blob==0.37.0
azure-storage-queue==0.37.0
boto3==1.4.4
connexion==1.1.16
google-cloud-pubsub==0.30.1
google-cloud-storage==1.6.0
python-swiftclient==3.4.0
python-keystoneclient==3.13.0
watchdog==0.8.3
git+git://github.com/ga4gh/data-object-schemas@master
oauthlib==3.0.0
requests_oauthlib==1.1.0
